Parser fait. Permet de chercher une solution. Cependant reste encore quelques bugs/choses à ajouter. 	modified:   ag41_project_sample1/Data.cpp 	modified:   ag41_project_sample1/Data.h 	modified:   ag41_project_sample1/Parser.cpp 	modified:   ag41_project_sample1/main.cpp

+++ ag41_project_sample1/Data.cpp
@@ -46,11 +46,11 @@ namespace Donnees
 		pair<Client*, Client*> p(c1, c2);
 		distanceClients[p] = dist;
 	}
-	int Data::getKTransport()
+	float Data::getKTransport()
 	{
 		return kTransport;
 	}
-	void Data::setKTransport(int k)
+	void Data::setKTransport(float k)
 	{
 		kTransport = k;
 	}
