+++ ag41_project_sample1/Data.h
@@ -19,7 +19,7 @@ namespace Donnees
 		map<Client*, int> distanceFournisseur;
 		vector<Client*> listeClient;
 		map<int, vector<Action*>* > sol;
-		int kTransport;
+		float kTransport;
 	public:
 		Data();
 		virtual ~Data();
@@ -30,8 +30,8 @@ namespace Donnees
 		void distanceClient(Client* c, int dist);
 		int distanceClient(Client* c1, Client* c2);
 		void distanceClient(Client* c1, Client* c2, int dist);
-		int getKTransport();
-		void setKTransport(int k);
+		float getKTransport();
+		void setKTransport(float k);
 
 
 	};
